+++ src/enumerables/__tests__/selectors.spec.js
@@ -18,5 +18,65 @@ test('enumerables returns correct state slice', () => {
       },
     }
   });
-  return expect(Immutable.is('foo-bar-bam', selectors.enumerables(state))).toBe(true);
+  expect(Immutable.is(
+    'foo-bar-bam',
+    selectors.enumerables(state)
+  )).toBe(true);
+});
+
+test("orderedEnumerable returns correct state slice", () => {
+  const state = Immutable.fromJS({
+    entities: {
+      enumerables: {
+        enumerables: {
+          states: {
+            NSW: "New South Wales",
+            QLD: "Queensland",
+            VIC: "Victoria"
+          },
+          countries: {
+            AU: "Australia",
+            NZ: "New Zealand",
+            GB: "England"
+          }
+        }
+      }
+    }
+  });
+  expect(Immutable.is(
+    state.getIn(['entities', 'enumerables', 'enumerables', 'countries']),
+    selectors.orderedEnumerable(state, 'countries').toMap()
+  )).toBe(true);
+});
+
+test("orderedEnumerable returns correct order", () => {
+  const state = Immutable.fromJS({
+    entities: {
+      enumerables: {
+        enumerables: {
+          states: {
+            NSW: "New South Wales",
+            QLD: "Queensland",
+            VIC: "Victoria"
+          },
+          countries: {
+            NZ: "New Zealand",
+            AU: "Australia",
+            ZB: "Zimbabwe",
+            GB: "England"
+          }
+        }
+      }
+    }
+  });
+  const shouldBe = Immutable.fromJS({
+    AU: "Australia",
+    GB: "England",
+    NZ: "New Zealand",
+    ZB: "Zimbabwe",
+  });
+  expect(Immutable.is(
+    selectors.orderedEnumerable(state, 'countries'),
+    shouldBe.toOrderedMap()
+  )).toBe(true);
 });