+++ src/enumerables/selectors.js
@@ -1,8 +1,12 @@
+import { Map } from 'immutable';
 import { SelectorFactory } from 'ethical-jobs-redux';
-import { createSelector } from 'reselect';
 
 export const fetching = SelectorFactory.create('enumerables', 'fetching');
 
 export const error = SelectorFactory.create('enumerables', 'error');
 
 export const enumerables = state => state.getIn(['entities', 'enumerables', 'enumerables']);
+
+export const orderedEnumerable = (state, enumerableKey) => state
+  .getIn(["entities", "enumerables", "enumerables", enumerableKey], Map())
+  .sortBy((value, key) => value);